remove functions superseded by ifcbkit
